backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
python-multipart==0.0.6
boto3==1.28.88
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
aiofiles==23.2.1
PyJWT==2.8.1
# AI Ana
lysis Dependencies
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
